services/postgres/requirements.txt
psycopg2-binary==2.9.9
Faker==19.13.0
werkzeug==3.0.1 